funnel/burnout_reset/requirements.txt
Flask>=2.0
APScheduler>=3.10
PyYAML>=6.0
requests>=2.28
SQLAlchemy>=2.0
Jinja2>=3.0
